+++ ak_library_management/__manifest__.py
@@ -2,15 +2,17 @@
     "name": "Library",
     "version": "18.0.1.12.0",
     "license": "Other proprietary",
-    "depends": ["sale_management","hr","stock","base_automation"],
-    "author":"Raj",
+    "depends": ["sale_management", "hr", "stock", "base_automation","website"],
+    "author": "Raj",
     "website": "https://www.aktivsoftware.com",
     "data": [
         "security/security.xml",
         "security/ir.model.access.csv",
+
         "report/custom_attachment_report_views.xml",
         "report/library_report_views.xml",
         "report/custom_library_report_views.xml",
+
         "views/library_book_tags_views.xml",
         "views/library_book_category.xml",
         "views/library_library_views.xml",
@@ -25,10 +27,15 @@
         "views/borrow_transaction_history_views.xml",
         "views/stock_warehouse_views.xml",
         "views/library_menu_views.xml",
+        "views/custom_web_menu.xml",
+        "views/custom_web_page.xml",
+
         "data/ir_sequence_data.xml",
         "data/ir_cron_data.xml",
         "data/ir_action_data.xml",
         "data/mail_template_data.xml",
+
+        "wizard/borrow_warning_wizard_views.xml",
         "wizard/sale_order_wizard_views.xml",
         "wizard/borrow_transaction_history_wizard_views.xml"
     ],