+++ ak_library_management/models/__init__.py
@@ -13,3 +13,4 @@ from . import res_partner
 from . import borrow_transaction_history
 from . import stock_warehouse
 from . import res_config_settings
+from . import res_company