+++ ak_library_management/models/res_company.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+from odoo import fields, models
+
+
+class ResCompany(models.Model):
+    _inherit = "res.company"
+
+    borrowing_limit = fields.Integer(
+        string="borrowing_limit"
+    )